backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.4.6",
    "aiohttp==3.11.12",
    "aiohttp-retry==2.9.1",
    "aiosignal==1.3.2",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anthropic==0.45.2",
    "anyio==4.8.0",
    "attrs==25.1.0",
    "backoff==2.2.1",
    "backtesting==0.6.1",
    "base58==2.1.1",
    "blinker==1.9.0",
    "bokeh==3.6.3",
    "cachetools==5.5.1",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "construct==2.10.68",
    "construct-typing==0.5.6",
    "contourpy==1.3.1",
    "decorator==4.4.2",
    "distro==1.9.0",
    "fastapi==0.109.0",
    "flask==3.1.0",
    "frozenlist==1.5.0",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "groq==0.16.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "idna==3.10",
    "imageio==2.37.0",
    "imageio-ffmpeg==0.6.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.2",
    "jiter==0.8.2",
    "joblib==1.4.2",
    "jsonalias==0.1.1",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "langdetect==1.0.9",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "moviepy==1.0.3",
    "multidict==6.1.0",
    "narwhals==1.26.0",
    "numpy==1.24.0",
    "numpy>=1.24.0 ; python_full_version >= '3.11'",
    "openai==1.59.5",
    "outcome==1.3.0.post0",
    "packaging==24.2",
    "pandas>=2.1.0",
    "pandas-ta>=0.3.14b0",
    "pillow==11.1.0",
    "proglog==0.1.10",
    "propcache==0.2.1",
    "protobuf==5.29.3",
    "pyarrow==19.0.0",
    "pycparser==2.22",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydeck==0.9.1",
    "pygments==2.19.1",
    "pyjwt==2.10.1",
    "pypdf2==3.0.0",
    "pysocks==1.7.1",
    "python-dateutil==2.8.2",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.6",
    "pytz==2025.1",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "requests==2.31.0",
    "rich==13.9.4",
    "rpds-py==0.22.3",
    "schedule==1.2.0",
    "selenium==4.28.1",
    "setuptools==75.8.0",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "solana==0.36.5",
    "solders==0.25.0",
    "sortedcontainers==2.4.0",
    "sounddevice==0.5.1",
    "starlette==0.35.1",
    "streamlit==1.42.0",
    "ta-lib==0.6.3",
    "tenacity==9.0.0",
    "termcolor==2.3.0",
    "toml==0.10.2",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "trio==0.28.0",
    "trio-websocket==0.11.1",
    "twilio==9.4.5",
    "typing-extensions==4.12.2",
    "tzdata==2025.1",
    "urllib3==2.3.0",
    "uvicorn==0.25.0",
    "watchdog==6.0.0",
    "wave==0.0.2",
    "webdriver-manager==4.0.2",
    "websocket-client==1.8.0",
    "websockets==14.2",
    "werkzeug==3.1.3",
    "wsproto==1.2.0",
    "xyzservices==2025.1.0",
    "yarl==1.18.3",
    "youtube-transcript-api==0.6.2",
    "yt-dlp==2025.1.26",
]
